feat: implement event controller and middleware for event retrieval with pagination
